+++ src/types.ts
@@ -74,6 +74,43 @@ interface FontObjectVariable {
 
 type FontObject = FontObjectV1 | FontObjectV2 | FontObjectVariable;
 
+// Variable axes - have to put here to prevent circular dependency
+// We manually add support for new axes to ensure they aren't breaking since Google likes to introduce new ways to use them
+const SUPPORTED_AXES_UPPER = [
+  "CASL",
+  "CRSV",
+  "FILL",
+  "GRAD",
+  "MONO",
+  "SOFT",
+  "WONK",
+  "XOPQ",
+  "XTRA",
+  "YOPQ",
+  "YTAS",
+  "YTDE",
+  "YTFI",
+  "YTLC",
+  "YTUC",
+] as const;
+const SUPPORTED_AXES_LOWER = ["ital", "opsz", "slnt", "wdth", "wght"] as const;
+const SUPPORTED_AXES = [
+  ...SUPPORTED_AXES_LOWER,
+  ...SUPPORTED_AXES_UPPER,
+] as const;
+
+type SupportedAxes = typeof SUPPORTED_AXES[number];
+
+const STANDARD_AXES = ["opsz", "slnt", "wdth", "wght"] as const;
+type StandardAxes = typeof STANDARD_AXES[number];
+
+const isAxesKey = (axesKey: string): axesKey is SupportedAxes =>
+  SUPPORTED_AXES.includes(axesKey as SupportedAxes);
+const isStandardAxesKey = (axesKey: string): axesKey is StandardAxes =>
+  STANDARD_AXES.includes(axesKey as StandardAxes);
+
+export { isAxesKey, isStandardAxesKey, STANDARD_AXES, SUPPORTED_AXES }
+
 export type {
   FontObject,
   FontObjectV1,
@@ -82,4 +119,6 @@ export type {
   FontObjectVariableDirect,
   FontVariants,
   FontVariantsVariable,
+  StandardAxes,
+  SupportedAxes
 };
